requirements.txt
opencv-python
numpy
torch
torchvision
pyyaml
flask
pyserial
